[Feature] Split PSR-16 into smaller interfaces for greater re-usability
_I partly expect this to maybe be closed as this change makes it harder to determin if some code actually implements PSR-16, but:_

Background for this was I found myself in position where I only needed single cache methods (get, set, delete) and clear()
for a internal in-memory class and was thinking there might be value in having a PSR interface to implement for it,
if someoen would find themselves in position where they'd like a different behavour.

I also see possible value in having a seperate interface for clear(), as it is something we need for any classes holding some
sort of objects in memory _(for cases we deem safe)_ to avoid re-loading the given object repeatedly.
